+++ Source/Types/Script.swift
@@ -7,8 +7,8 @@
 import Foundation
 
 public enum ScriptHashType: String, Codable {
-    case data = "Data"
-    case type = "Type"
+    case data
+    case type
 }
 
 public struct Script: Codable, Param {
@@ -22,7 +22,9 @@ public struct Script: Codable, Param {
         case hashType = "hash_type"
     }
 
+    // Before serialization is implemented _compute_script_hash RPC should be used instead.
     public var hash: String {
+        #warning("This needs re-implementation when serialization is applied.")
         var bytes = [UInt8]()
         bytes.append(contentsOf: Data(hex: codeHash).bytes)
         bytes.append(hashType == .data ? 0x0 : 0x1)