+++ Tests/API/APIMockingTests.swift
@@ -59,7 +59,7 @@ class APIMockingTests: XCTestCase {
     }
 
     func testGetHeader() throws {
-        let hash = "0xba0d878d2c3711d38b5ddc2bc917312ca3898cad98457cc7960e28ec31f26e7f"
+        let hash = "0x205130d98f10bccc4a5d4643fe6dae6cb8c17ff9304f6de233efce50aa525ef3"
         let result = try getClient(json: "header").getHeader(blockHash: hash)
         XCTAssertNotNil(result)
         XCTAssertEqual(result.hash, hash)
@@ -73,10 +73,7 @@ class APIMockingTests: XCTestCase {
     }
 
     func testGetLiveCell() throws {
-        let outPoint = OutPoint(
-            blockHash: "0x4c2f8ba5f5a0104eaf84fcbb16af4b0e7ca2f2fdb076e748d54ef876d085d49e",
-            cell: CellOutPoint(txHash: "0xda27a795e14067e18a0cfc0951571aaca47bd9851adc9bf5baa22cf27c8bcde8", index: "0")
-        )
+        let outPoint = OutPoint(txHash: "0xda27a795e14067e18a0cfc0951571aaca47bd9851adc9bf5baa22cf27c8bcde8", index: "0")
         let result = try getClient(json: "liveCellEmpty").getLiveCell(outPoint: outPoint)
         XCTAssertNotNil(result)
     }
@@ -134,9 +131,23 @@ class APIMockingTests: XCTestCase {
     }
 
     func testComputeTransactionHash() throws {
-        let tx = Transaction()
+        let tx = Transaction(
+            cellDeps: [CellDep(outPoint: OutPoint(txHash: "0xca457e8f2babbced0321daa535d5d357f554ff601a164c3ce76b547fd5ad2452", index: "0"), depType: .code)],
+            headerDeps: ["0xb877a54927a2cc36f455533160f65025c0066a3a371a81be2a9bb75554abcc43"],
+            inputs: [CellInput(previousOutput: OutPoint(txHash: "0xd134dabdb4686abd4e25b0fc95d5b7da5e14efc87e99bd34965571725f6370e3", index: "0"), since: "0")],
+            outputs: [CellOutput(capacity: "100000000000", lock: Script(args: [], codeHash: "0x28e83a1277d48add8e72fadaa9248559e1b632bab2bd60b27955ebc4c03800a5", hashType: .data))],
+            outputsData: ["0x"]
+        )
         let result = try? getClient(json: "computeTransactionHash").computeTransactionHash(transaction: tx)
         XCTAssertNotNil(result)
+        XCTAssertEqual("0x45746b0f98b11fa67c88b66ac61044238af6f1e69f0006c690998fbe0cc96878", result)
+    }
+
+    func testComputeScriptHash() throws {
+        let script = Script(args: [], codeHash: "0x28e83a1277d48add8e72fadaa9248559e1b632bab2bd60b27955ebc4c03800a5", hashType: .data)
+        let result = try? getClient(json: "computeScriptHash").computeScriptHash(script: script)
+        XCTAssertNotNil(result)
+        XCTAssertEqual("0xd8753dd87c7dd293d9b64d4ca20d77bb8e5f2d92bf08234b026e2d8b1b00e7e9", result)
     }
     
     func testDryRunTransaction() throws {